- Remove `EventDispatcher` files - Rename and restructure renderer files - Add `BufferLayout` class - Comment out test code in `start.cpp` - Update CMakeLists and `.clang-tidy`
Removed obsolete `EventDispatcher` files and refactored renderer components to better-organized directories. Introduced a new `BufferLayout` class to handle buffer schemas. This commit improves the project's structure and future maintainability. This commit message was generated by AI.
