+++ .clang-tidy
@@ -44,7 +44,6 @@ modernize-*,
 google-build-namespaces,
 google-default-arguments,
 google-build-explicit-make-pair,
-google-readability-namespace-comments,
 google-global-names-in-headers,
 google-runtime-int,
 google-runtime-operator,
@@ -58,7 +57,6 @@ bugprone-*,
 - bugprone-easily-swappable-parameters, 
 llvm-twine-local,
 llvm-include-order,
-llvm-namespace-comment,
 performance-*,
 -misc-include-cleaner,
 cppcoreguidelines-*,
