+++ CMakeLists.txt
@@ -37,6 +37,7 @@ endif()
 #       ### Sub Directories ###
 add_subdirectory(vendor/GLAD)
 add_subdirectory(vendor/ImGui)
+add_subdirectory(vendor/glm)
 add_subdirectory(source/Ayla)
 add_subdirectory(example-user-project)
