+++ source/Ayla/Core/Layers/Layer.h
@@ -4,7 +4,6 @@
 
 #pragma once
 #include "Ayla/Events/Event.h"
-#include "Ayla/Events/EventDispatcher.h"
