+++ source/Ayla/ECS/Entity.h
@@ -4,9 +4,11 @@
 
 #pragma once
 
-
 #include "../../../vendor/glm/glm/vec2.hpp"
 #include "../../../vendor/glm/glm/vec4.hpp"
+
+#include "Ayla/Renderer/Abstractions/IndexBufferObject.h"
+#include "Ayla/Renderer/Abstractions/VertexBuffer.h"
 
 // TEMP
 namespace Ayla::ECS {
@@ -19,11 +21,14 @@ namespace Ayla::ECS {
         Entity(glm::vec2 position, glm::vec4 color);
         ~Entity() = default;
 
+
         glm::vec2 m_Position;
         glm::vec4 m_Color;
+        glm::vec2 m_Vertices[4];
 
     private:
-
+        Renderer::IndexBufferObject m_IndexBuffer = Renderer::IndexBufferObject(,);
+        Renderer::VertexBuffer m_VertexBuffer = Renderer::VertexBuffer(,);
 
     };
