--- source/Ayla/Events/EventDispatcher.cpp
@@ -1,10 +1,0 @@
-//
-// Created by Andrew Fagan on 4/18/24.
-//
-#include "Ayla/Events/EventDispatcher.h"
-
-namespace Ayla::EventManagement {
-
-
-
-} // namespace Ayla::EventManagement
