+++ source/Ayla/Renderer/Abstractions/BufferLayout.cpp
@@ -1,0 +1,5 @@
+//
+// Created by Andrew Fagan on 7/24/24.
+//
+
+#include "BufferLayout.h"
